Corrected separation error and switched to ramp extenion
I corrected an error in the separation of the RNA ramps/steps.  Now
should be corrected.  I also now load the "extension" towards of the
surface of the AR ramp.  This will be useful for to make sure the AR
ramps are lining up with the RNA ramp/steps AND for work energy analysis
later.