Initial RNA Viewer procs and settings
I'm now making an application to just view the RNA pulling program data.
The goals of this program are to add features, such as splitting
approach and retract ramps, and add the ability to color the graphs by
the HMM state identification.  I also want to be able to use this
program while running the RNA pulling program on the cypher, so I can
see graphs in terms of force and extension (not just voltages)